extend session-key validation to bookmark + reader-settings paths (#608)
The same stale-session_key FK-violation pattern existed in two more
places that also write rows whose session FK can be stale:

- BookmarkAuthMixin.get_bookmark_auth_filter — feeds session_id into
  Bookmark.objects.bulk_create / bulk_update.
- ReaderSettingsBaseView._get_bookmark_auth_filter — feeds session_id
  into SettingsReader.objects.create.

Both used the old `if not session.session_key: save()` pattern that
trusts the cookie. Hoist the validated _ensure_session_key helper
from SettingsBaseView up to AuthMixin so every auth-aware view shares
one implementation, and switch both call sites to it. BookmarkAuthMixin
now extends AuthMixin to inherit the helper. BookmarkFilterMixin is
unchanged — it's read-only (filter Q only) and a missing session
correctly returns no rows.
